[WIP] ACM-2071 - Collect status conditions with Configurable Collection
